[#5](status:Fixed user:schmidtw) Added support for up to 999 songs per disc, and up to 65 discs.

+++ library/database/src/add_song.c
@@ -9,7 +9,7 @@ ll_ir_t scrubber(ll_node_t *node, volatile void *user_data);
 
 song_node_t * add_song_to_group( group_node_t * group,
         const char * artist, const char * album,
-        const char * song, const uint8_t track_number,
+        const char * song, const uint16_t track_number,
         media_command_fn_t command_fn,
         media_play_fn_t play_fn,
         const char * file_location )
